Cambia delete de hotel

+++ Back/src/repositories/hotel.repository.ts
@@ -4,7 +4,6 @@ import { pool } from '../shared/conn.js';
 import { ResultSetHeader, RowDataPacket } from 'mysql2';
 
 export class HotelRepository implements Repository<Hotel> {
-  
   public async findAll(): Promise<Hotel[] | undefined> {
     const [hoteles] = await pool.query<RowDataPacket[]>(
       'SELECT * FROM hoteles'
@@ -22,7 +21,7 @@ export class HotelRepository implements Repository<Hotel> {
     if (hoteles_raw.length == 0) {
       return undefined;
     }
-    
+
     const hotel = hoteles_raw[0] as Hotel;
 
     return hotel;
@@ -31,7 +30,15 @@ export class HotelRepository implements Repository<Hotel> {
   public async save(item: Hotel): Promise<Hotel | undefined> {
     const [result] = await pool.query<[RowDataPacket[], ResultSetHeader]>(
       'INSERT INTO hoteles (nombre, direccion, descripcion, telefono, email, estrellas, id_ciudad) VALUES (?, ?, ?, ?, ?, ?, ?)',
-      [item.nombre, item.direccion, item.descripcion, item.telefono, item.email, item.estrellas, item.id_ciudad]
+      [
+        item.nombre,
+        item.direccion,
+        item.descripcion,
+        item.telefono,
+        item.email,
+        item.estrellas,
+        item.id_ciudad,
+      ]
     );
     const affectedRows = (result as any).affectedRows;
     if (affectedRows == 1) {
@@ -41,14 +48,41 @@ export class HotelRepository implements Repository<Hotel> {
     }
   }
 
-  public async update(item: { id: string }, hotel: Hotel): Promise<Hotel | undefined> {
+  public async update(
+    item: { id: string },
+    hotel: Hotel
+  ): Promise<Hotel | undefined> {
     const id = Number.parseInt(item.id);
-    await pool.query('UPDATE hoteles SET nombre = ?, direccion = ?, descripcion = ?, telefono = ?, email = ?, estrellas = ?, id_ciudad = ? WHERE id = ?', [hotel.nombre, hotel.direccion, hotel.descripcion, hotel.telefono, hotel.email, hotel.estrellas, hotel.id_ciudad, id]);
+    await pool.query(
+      'UPDATE hoteles SET nombre = ?, direccion = ?, descripcion = ?, telefono = ?, email = ?, estrellas = ?, id_ciudad = ? WHERE id = ?',
+      [
+        hotel.nombre,
+        hotel.direccion,
+        hotel.descripcion,
+        hotel.telefono,
+        hotel.email,
+        hotel.estrellas,
+        hotel.id_ciudad,
+        id,
+      ]
+    );
     return hotel;
   }
-  
+
   public async remove(item: { id: string }): Promise<void> {
     const id = Number.parseInt(item.id);
+
+    const [estadias] = await pool.query<RowDataPacket[]>(
+      'SELECT * FROM estadias WHERE id_hotel = ?',
+      [id]
+    );
+
+    if (estadias.length > 0) {
+      throw new Error(
+        'No se puede eliminar el hotel porque está asociado a una o más estadías.'
+      );
+    }
+
     await pool.query('DELETE FROM hoteles WHERE id = ?', [id]);
   }
 }
